General clean up and imported ammo asset + implemented it
