+++ Example2-Unity Hybrid-Survival/Assets/AddressableAssetsData/AssetGroups/Default Local Group.asset
@@ -71,7 +71,7 @@ MonoBehaviour:
     m_SerializedLabels: []
     m_MainAsset: {fileID: 0}
     m_TargetAsset: {fileID: 0}
-  - m_GUID: 9e6c066266bf45a4382b4e1f81fbe8cf
+  - m_GUID: e93a28b875cc5eb41b51bd517c2ac791
     m_Address: AmmoCrate
     m_ReadOnly: 0
     m_SerializedLabels: []
