+++ Example2-Unity Hybrid-Survival/UserSettings/EditorUserSettings.asset
@@ -6,6 +6,9 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedScenePath-0:
+      value: 224247031146467a000303001f3b10311b1a1708233068252320092a
+      flags: 0
+    RecentlyUsedScenePath-1:
       value: 22424703114646680e0b0227036c73150012146a7c7968252320092a
       flags: 0
     vcSharedLogLevel:
